+++ NEWS.md
@@ -8,6 +8,7 @@
 
 + `tradeCycle` includes number of business days in trade cycles based on Nymex calendar.
 + Added Canadian crude ICE contracts expiry in `expiry_table`.
++ `cma` contains metadata for computation of CME WTI Calendar Month Average swap.
 
 # RTL 1.3.2